// src/services/models/memeDetailResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Swipe Pad API
 * OpenAPI spec version: 0.0.12
 */
import type { MemeDetailResponseStatus } from "./memeDetailResponseStatus";
import type { UserResponse } from "./userResponse";

export interface MemeDetailResponse {
  readonly id: number;
  readonly created_at: string;
  readonly updated_at: string;
  readonly status: MemeDetailResponseStatus;
  readonly tokenAddress: string;
  readonly submissionId: string;
  readonly quoteAddress: string;
  readonly creator: string;
  readonly name: string;
  readonly symbol: string;
  readonly decimals: string;
  readonly tags: readonly string[];
  readonly caption: string;
  readonly image: string;
  readonly banner: string;
  readonly description: string;
  readonly x: string;
  readonly telegram: string;
  readonly discord: string;
  readonly github: string;
  readonly website: string;
  readonly guarded: boolean;
  readonly isNoFee: boolean;
  readonly tokenOffset: string;
  readonly quoteOffset: string;
  readonly totalSupply: string;
  readonly saleAmount: string;
  readonly virtualQuoteReserve: string;
  readonly virtualTokenReserve: string;
  readonly pairAddress: string;
  readonly createdTime: number;
  readonly graduatedTime: number;
  readonly volume24hs: string;
  readonly currentPriceByUsd: string;
  readonly price24hChange: string;
  readonly volumeBuyByQuote: string;
  readonly kingOfTheHillTime: number;
  readonly totalStars: number;
  readonly totalLikes: number;
  readonly totalComments: number;
  readonly isUserLike: boolean;
  readonly listProgress: string;
  readonly kingOfTheHillProgress: string;
  readonly marketCap: string;
  readonly userInfo: UserResponse;
}