// src/services/models/memeDetailResponseStatus.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Swipe Pad API
 * OpenAPI spec version: 0.0.12
 */

export type MemeDetailResponseStatus =
  (typeof MemeDetailResponseStatus)[keyof typeof MemeDetailResponseStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MemeDetailResponseStatus = {
  active: "active",
  hidden: "hidden",
  ended: "ended",
  completed: "completed",
} as const;